requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
blinker==1.9.0
breadability==0.1.20
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cmudict==1.0.32
cohere==5.15.0
colorama==0.4.6
deep-translator==1.11.4
distro==1.9.0
docopt==0.6.2
fastavro==1.11.1
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
google-api-core==2.25.0
google-auth==2.40.3
google-genai==1.19.0
googleapis-common-protos==1.70.0
groq==0.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
language_tool_python==2.9.3
lxml==5.4.0
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.2.6
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.5
pydantic_core==2.33.2
pyphen==0.17.2
python-dotenv==1.1.0
PyYAML==6.0.2
rake-nltk==1.0.6
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
Sastrawi==1.0.1
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
sumy==0.11.0
textstat==0.7.7
tokenizers==0.21.1
toml==0.10.2
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.22.0
